[server.2] Get initial Google Cloud implementation in place (#134)
## Summary:
This migrates the remainder of the `shared` portion of Khan/render-gateway into Wonder Stuff, building the first pieces of a Google Cloud-supporting layer.

I think I still need a base trace implementation in Wonder Stuff Server. That way, we can have drop in replacements for code and easily switch platforms. I'll make that tweak in the next PR as this was a pretty big one and that should close out this phase.

After that, I'll be working on the `gateway` part of the Khan/render-gateway repo. I think ultimately, the stuff in `ka-shared` will actually be things we would move to the implementation in the webapp service or a combination of that and some helpers in one of the existing server packages here in Wonder Stuff instead. So, I'll work on that last as the webapp render-gateway gets reimplemented using these packages.

Issue: FEI-4004

## Test plan:
`yarn test`

Author: somewhatabstract

Reviewers: somewhatabstract, jeresig

Required Reviewers:

Approved By: jeresig

Checks: ⌛ Analyze (javascript), ⌛ Prime node_modules cache for primary configuration (ubuntu-latest, 14.x), ⏭  dependabot

Pull Request URL: https://github.com/Khan/wonder-stuff/pull/134