Mark demo project as non-publishable (#1139)
## Summary:

The most recent release failed near the end because it tried to publish the `demo` project. This PR changes the demo project to be explicitly private. This will help changesets know not to try to publish it. 

This didn't happen in the demo project in earlier releases because it wasn't in the Yarn workspace. With pnpm, we can add it to the workspace, but that caused changesets to try to publish it.  By marking it as `private: true`, we'll block that in the future. 

<img width="1312" alt="image" src="https://github.com/user-attachments/assets/f2f38370-f349-47dd-b713-e4605a3405ec" />

Issue: FEI-6300

## Test plan:

We'll need to land this and then release the Version Packages PR that's created by it.

Author: jeremywiebe

Reviewers: somewhatabstract

Required Reviewers:

Approved By: somewhatabstract

Checks: ✅ 14 checks were successful, ⏭️  3 checks have been skipped

Pull Request URL: https://github.com/Khan/wonder-stuff/pull/1139