Add eslint-plugin-khan to devDependencies so workspace:* version works (#1225)
## Summary:

On the [last release](https://github.com/Khan/wonder-stuff/actions/runs/19450157527/job/55653123539) from this repo, the `@khanacademy/eslint-config` failed to publish because it uses a `workspace:*` version for the `@khanacademy/eslint-plugin` peerDependency. It specifies `@khanacademy/eslint-plugin` as a peer because it's not technically dependeded on, but does reference the plugin in its config (maybe that's a sign that `@khanacademy/eslint-plugin` _should_ be a regular dependency?).  Other plugins that are referenced in the config are referenced as peer dependencies so I think this is correct.

To fix this for now, I've added the plugin to `devDependencies`. When that's done, `pnpm` properly parses and updates everything when it publishes. 

The error:

```
🦋  error an error occurred while publishing @khanacademy/eslint-config: ERR_PNPM_CANNOT_RESOLVE_WORKSPACE_PROTOCOL undefined
🦋  error {
🦋  error   "error": {
🦋  error     "code": "ERR_PNPM_CANNOT_RESOLVE_WORKSPACE_PROTOCOL",
🦋  error     "message": "Cannot resolve workspace protocol of dependency \"@khanacademy/eslint-plugin\" because this dependency is not installed. Try running \"pnpm install\"."
🦋  error   }
🦋  error }
🦋  error
```

Issue: LEMS-3681

## Test plan:

Land this PR and do another release. :sigh:

Author: jeremywiebe

Reviewers: jeremywiebe, jeresig, somewhatabstract, kevinb-khan

Required Reviewers:

Approved By: jeresig

Checks: ✅ 6 checks were successful, ❌ 1 check failed, ⏭️  2 checks have been skipped

Pull Request URL: https://github.com/Khan/wonder-stuff/pull/1225